requirements.txt
pandas
openpyxl
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
python-multipart
openpyxl